+++ wdtk-client/src/main/java/org/wikidata/wdtk/client/RdfSerializationAction.java
@@ -282,6 +282,9 @@ public class RdfSerializationAction extends DumpProcessingOutputAction {
 		String message = "Finished serialization of "
 				+ this.serializer.getTripleCount() + " RDF triples in file "
 				+ this.insertDumpInformation(this.outputDestination);
+		if (this.compressionType != COMPRESS_NONE) {
+			message += "." + this.compressionType;
+		}
 		return message;
 	}