+++ wdtk-util/src/main/java/org/wikidata/wdtk/util/WebResourceFetcherImpl.java
@@ -22,6 +22,7 @@ package org.wikidata.wdtk.util;
 
 import java.io.IOException;
 import java.io.InputStream;
+import java.net.Proxy;
 import java.net.URL;
 import java.net.URLConnection;
 
@@ -36,9 +37,42 @@ public class WebResourceFetcherImpl implements WebResourceFetcher {
 	protected static String userAgent = "Wikidata Toolkit; Java "
 			+ System.getProperty("java.version");
 
+	protected static Proxy proxy = null;
+
 	/**
-	 * Returns the string that will be used to ideintify the user agent on all
-	 * requests made by Wikidata Toolkit.
+	 * Returns the proxy that will be used for all requests made by Wikidata
+	 * Toolkit.
+	 * 
+	 * @return the proxy represented as java object
+	 */
+	public static Proxy getProxy() {
+		return proxy;
+	}
+
+	/**
+	 * Sets the proxy that will be used for alle requests made by Wikidata
+	 * Toolkit. This should be set in own tools based on Wikidata Toolkit
+	 * esp. when making large amounts of requests.
+	 * 
+	 * @param proxy
+	 *                the proxy represented as java object
+	 */
+	public static void setProxy(Proxy proxy) {
+		WebResourceFetcherImpl.proxy = proxy;
+	}
+
+	/**
+	 * Checks whether a proxy is set.
+	 * 
+	 * @return True if a proxy is set, false, if there isn't set any proxy.
+	 */
+	public static boolean hasProxy() {
+		return (proxy != null);
+	}
+
+	/**
+	 * Returns the string that will be used to identify the user agent on
+	 * all requests made by Wikidata Toolkit.
 	 *
 	 * @return the user agent string
 	 */
@@ -47,12 +81,12 @@ public class WebResourceFetcherImpl implements WebResourceFetcher {
 	}
 
 	/**
-	 * Sets the string that will be used to ideintify the user agent on all
-	 * requests made by Wikidata Toolkit. This should be set in own tools based
-	 * on Wikidata Toolkit esp. when making large amounts of requests.
+	 * Sets the string that will be used to identify the user agent on all
+	 * requests made by Wikidata Toolkit. This should be set in own tools
+	 * based on Wikidata Toolkit esp. when making large amounts of requests.
 	 *
 	 * @param userAgent
-	 *            the user agent string
+	 *                the user agent string
 	 */
 	public static void setUserAgent(String userAgent) {
 		WebResourceFetcherImpl.userAgent = userAgent;
@@ -62,7 +96,12 @@ public class WebResourceFetcherImpl implements WebResourceFetcher {
 	public InputStream getInputStreamForUrl(String urlString)
 			throws IOException {
 		URL url = new URL(urlString);
-		URLConnection urlConnection = url.openConnection();
+		URLConnection urlConnection;
+		if (hasProxy()) {
+			urlConnection = url.openConnection(proxy);
+		} else {
+			urlConnection = url.openConnection();
+		}
 		urlConnection.setRequestProperty("User-Agent", userAgent);
 
 		return urlConnection.getInputStream();